Changed variable type. Added Scale01Clamped func.

+++ src/Particle/ParticleFwd.h
@@ -12,9 +12,9 @@ namespace mono
     struct ParticlePoolComponent;
     struct ParticleEmitterComponent;
 
-    using ParticleGenerator = std::function<void (const math::Vector& position, ParticlePoolComponent& pool, size_t index)>;
-    using ParticleUpdater   = std::function<void (ParticlePoolComponent& pool, size_t count, uint32_t delta_ms)>;
+    using ParticleGenerator = std::function<void (const math::Vector& position, ParticlePoolComponent& pool, uint32_t index)>;
+    using ParticleUpdater   = std::function<void (ParticlePoolComponent& pool, uint32_t count, uint32_t delta_ms)>;
 
-    void DefaultGenerator(const math::Vector& position, ParticlePoolComponent& pool, size_t index);
-    void DefaultUpdater(ParticlePoolComponent& pool, size_t count, uint32_t delta_ms);
+    void DefaultGenerator(const math::Vector& position, ParticlePoolComponent& pool, uint32_t index);
+    void DefaultUpdater(ParticlePoolComponent& pool, uint32_t count, uint32_t delta_ms);
 }
